+++ lizard_map/media/lizard_map/lizard_map.js
@@ -515,17 +515,30 @@ function setUpLegendEdit() {
 }
 
 
+/*
+Sends current extent and name of visible base layer.
+*/
 function mapSaveLocation() {
-    var url, extent;
+    var url, extent, visible_base_layer_name, i;
     url = $("#lizard-map-wms").attr("data-save-location-url");
     extent = map.getExtent();
+    visible_base_layer_name = "";
+
+    // Find the active base layer.
+    for (i = 0; i < map.layers.length; i = i + 1) {
+        if (map.layers[i].isBaseLayer && map.layers[i].visibility) {
+            visible_base_layer_name = map.layers[i].name;
+            break;
+        }
+    }
     $.ajax({
         type: 'POST',
         url: url,
         data: {bottom: extent.bottom,
                left: extent.left,
                right: extent.right,
-               top: extent.top},
+               top: extent.top,
+               base_layer_name: visible_base_layer_name},
         async: false,
         success: function () {}
     });